refactor: split AST toolchain helpers
