+++ tests/test_dumper_unit.py
@@ -25,7 +25,6 @@ from abicheck.dumper import (
     _resolve_debug_metadata,
     _safe_mtime,
     _safe_size,
-    _tool_identity,
     _vt_sort_key,
 )
 from abicheck.model import Visibility
@@ -165,42 +164,6 @@ class TestCacheKey:
         k1 = _cache_key([h], [], "c++")
         k2 = _cache_key([h], [], "cc")
         assert k1 != k2
-
-    def test_frontend_binary_identity_is_part_of_key(self, tmp_path):
-        h = tmp_path / "foo.h"
-        h.write_text("int x;", encoding="utf-8")
-        k1 = _cache_key(
-            [h], [], "c++", frontend_identity="castxml 0.6; clang 17"
-        )
-        k2 = _cache_key(
-            [h], [], "c++", frontend_identity="castxml 0.6; clang 21"
-        )
-        assert k1 != k2
-
-    def test_host_compiler_identity_is_part_of_key(self, tmp_path):
-        h = tmp_path / "foo.h"
-        h.write_text("int x;", encoding="utf-8")
-        k1 = _cache_key([h], [], "c++", compiler_identity="gcc 12")
-        k2 = _cache_key([h], [], "c++", compiler_identity="gcc 13")
-        assert k1 != k2
-
-    def test_tool_identity_changes_for_replaced_same_version_binary(self, tmp_path):
-        tool = tmp_path / "tool"
-        tool.write_text("#!/bin/sh\necho tool-version-1\n# A\n", encoding="utf-8")
-        tool.chmod(0o755)
-        identity_a = _tool_identity(str(tool))
-
-        replacement = tmp_path / "replacement"
-        replacement.write_text(
-            "#!/bin/sh\necho tool-version-1\n# B\n", encoding="utf-8"
-        )
-        replacement.chmod(0o755)
-        replacement.replace(tool)
-        identity_b = _tool_identity(str(tool))
-
-        assert identity_a != identity_b
-        assert "sha256=" in identity_a
-        assert "sha256=" in identity_b
 
     def test_with_include_dirs(self, tmp_path):
         h = tmp_path / "foo.h"
